MOBI-721 - Tax & discount mismatch with the reference order

+++ src/Tool/Def/Format.php
@@ -11,14 +11,12 @@ use Praxigento\Core\Config as Cfg;
 class Format
     implements \Praxigento\Core\Tool\IFormat
 {
-    /** @inheritdoc */
     public function dateTimeForDb(\DateTime $dt)
     {
         $result = $dt->format(Cfg::FORMAT_DATETIME);
         return $result;
     }
 
-    /** @inheritdoc */
     public function roundBonus($val, $precision = 2)
     {
         $round = round($val, $precision, PHP_ROUND_HALF_UP);
@@ -26,10 +24,9 @@ class Format
         return $result;
     }
 
-    /** @inheritdoc */
     public function toNumber($val, $precision = 2, $decimal = '.', $group = '')
     {
-        $result = (float) number_format($val, $precision, $decimal, $group);
+        $result = number_format($val, $precision, $decimal, $group);
         return $result;
     }
 }
